examples/python-examples/jaeger-ot-demo/requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
contextlib2==21.6.0
Deprecated==1.2.18
Flask==3.1.1
googleapis-common-protos==1.70.0
grpcio==1.72.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
jaeger-client==4.8.0
Jinja2==3.1.6
MarkupSafe==3.0.2
opentelemetry-api==1.33.1
opentelemetry-distro==0.54b1
opentelemetry-exporter-otlp==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asyncio==0.54b1
opentelemetry-instrumentation-click==0.54b1
opentelemetry-instrumentation-dbapi==0.54b1
opentelemetry-instrumentation-flask==0.54b1
opentelemetry-instrumentation-grpc==0.54b1
opentelemetry-instrumentation-jinja2==0.54b1
opentelemetry-instrumentation-logging==0.54b1
opentelemetry-instrumentation-requests==0.54b1
opentelemetry-instrumentation-sqlite3==0.54b1
opentelemetry-instrumentation-threading==0.54b1
opentelemetry-instrumentation-tornado==0.54b1
opentelemetry-instrumentation-urllib==0.54b1
opentelemetry-instrumentation-urllib3==0.54b1
opentelemetry-instrumentation-wsgi==0.54b1
opentelemetry-opentracing-shim==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
opentracing==2.4.0
packaging==25.0
protobuf
requests==2.32.5
six==1.17.0
threadloop==1.0.2
thrift==0.22.0
tornado==6.5
typing_extensions==4.14.0
urllib3==2.6.3
Werkzeug==3.1.4
wrapt==1.17.2
zipp==3.22.0
